pyproject.toml
[tool.poetry]
name = "quizio-backend"
version = "0.1.0"
description = ""
authors = ["kaushu42 <kaushu42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
channels = "^4.2.0"
uvicorn = {extras = ["standard"], version = "^0.32.1"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
